Implemented Clean file structure and lint rules added
