+++ lib/src/domain/entity/genre.dart
